+++ packages/core/src/components/TextArea/TextArea.tsx
@@ -1,10 +1,12 @@
-import React, { forwardRef } from "react";
+import React, { forwardRef, useCallback, useMemo, useState } from "react";
 import cx from "classnames";
 import { getTestId } from "../../tests/test-ids-utils";
 import { ComponentDefaultTestId } from "../../tests/constants";
 import styles from "./TextArea.module.scss";
 import { TextAreaProps, TextAreaSize } from "./TextArea.types";
 import Text from "../Text/Text";
+import { Flex } from "../Flex";
+import { HiddenText } from "../HiddenText";
 
 const DEFAULT_ROWS: Record<TextAreaSize, number> = {
   small: 3,
@@ -26,20 +28,42 @@ const TextArea = forwardRef(
       disabled,
       readOnly,
       required,
+      maxLength,
+      allowExceedingMaxLength,
+      onChange,
+      value,
       resize = true,
+      showCharCount = false,
       ...rest
     }: TextAreaProps,
     ref: React.ForwardedRef<HTMLTextAreaElement>
   ) => {
     const numRows = rows || DEFAULT_ROWS[size];
     const helpTextId = helpText && `${id}-help-text`;
+    const allowExceedingMaxLengthTextId = allowExceedingMaxLength && `${id}-allow-exceeding-max-length`;
+    const isErrorState = error || (maxLength && value?.length > maxLength);
+
+    const ariaDescribedby = useMemo(
+      () => [helpTextId, allowExceedingMaxLengthTextId].filter(id => !!id).join(" ") || undefined,
+      [helpTextId, allowExceedingMaxLengthTextId]
+    );
+
+    const [characterCount, setCharacterCount] = useState(value?.length || 0);
+
+    const handleOnChange = useCallback(
+      (event: React.ChangeEvent<HTMLTextAreaElement>) => {
+        setCharacterCount(event.target.value.length);
+        onChange?.(event);
+      },
+      [onChange]
+    );
 
     return (
       <div
         className={cx(
           styles.textAreaWrapper,
           {
-            [styles.error]: error,
+            [styles.error]: isErrorState,
             [styles.success]: success,
             [styles.disabled]: disabled,
             [styles.readOnly]: readOnly
@@ -56,20 +80,34 @@ const TextArea = forwardRef(
         <textarea
           {...rest}
           id={id}
+          maxLength={typeof maxLength === "number" && !allowExceedingMaxLength ? maxLength : undefined}
           ref={ref}
+          value={value}
           disabled={disabled}
           readOnly={readOnly}
           required={required}
           rows={numRows}
           className={cx(styles.textArea, [styles[size]], { [styles.resize]: resize })}
           aria-invalid={error}
-          aria-describedby={helpTextId ?? undefined}
+          aria-describedby={ariaDescribedby}
+          onChange={handleOnChange}
         />
-        {helpText && (
-          <Text className={cx(styles.helpText)} color="inherit" id={helpTextId}>
-            {helpText}
-          </Text>
-        )}
+        <Flex gap="xs" justify="space-between" className={cx(styles.subTextContainer)}>
+          {helpText && (
+            <Text className={cx(styles.helpText)} color="inherit" id={helpTextId}>
+              {helpText}
+            </Text>
+          )}
+          {showCharCount && (
+            <>
+              <Text className={styles.limitText}>
+                {characterCount}
+                {typeof maxLength === "number" && `/${maxLength}`}
+              </Text>
+              <HiddenText id={allowExceedingMaxLengthTextId} text={`Maximum of ${maxLength} characters`} />
+            </>
+          )}
+        </Flex>
       </div>
     );
   }